Check that DB is upgraded to latest Alembic version
This prevents runtime errors such as "table does not exist", because the
Pyramid app will not even start in case the DB is not upgraded to latest
Alembic version. We had this in many of our apps for years and it's time
to put it in a library instead of copy/pasting across projects.

The downside is that `pyramid_deferred_sqla` now expects to find
`etc/alembic.ini`. But that's also a good things, since now all apps will
have it in the same place.